Changed unsigned int -> size_t

+++ ft_memchr.c
@@ -6,13 +6,15 @@
 /*   By: pnguyen- <pnguyen-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/06 16:42:03 by pnguyen-          #+#    #+#             */
-/*   Updated: 2023/11/09 19:13:01 by pnguyen-         ###   ########.fr       */
+/*   Updated: 2023/11/11 16:36:36 by pnguyen-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	*ft_memchr(const void *s, int c, unsigned int n)
+#include <stddef.h>
+
+void	*ft_memchr(const void *s, int c, size_t n)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	while (i < n)
