+++ ft_strlen.c
@@ -6,13 +6,15 @@
 /*   By: pnguyen- <pnguyen-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/06 12:23:55 by pnguyen-          #+#    #+#             */
-/*   Updated: 2023/11/09 19:17:29 by pnguyen-         ###   ########.fr       */
+/*   Updated: 2023/11/11 16:10:50 by pnguyen-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-unsigned int	ft_strlen(const char s[])
+#include <stddef.h>
+
+size_t	ft_strlen(const char s[])
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i] != '\0')
